+++ src/Catalog/CatalogServiceInterface.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 
 namespace Wizaplace\SDK\Catalog;
 
+use Psr\Http\Message\ResponseInterface;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 
@@ -160,4 +161,11 @@ interface CatalogServiceInterface
      * @return ProductAttributeValue|null
      */
     public function getBrandFromProduct(Product $product): ?ProductAttributeValue;
+
+    /**
+     * @param string $attachmentId
+     *
+     * @return ResponseInterface
+     */
+    public function getProductAttachment(string $attachmentId): ResponseInterface;
 }