+++ CHANGELOG.md
@@ -5,6 +5,24 @@
 ### New features
 
 - Added `Wizaplace\SDK\Catalog\Product::isMVP` to know if a product was a Multi-Vendor product or not
+- Added `Wizaplace\SDK\Basket\Payment::getExternalReference`
+- Added `Wizaplace\SDK\Order\Payment::getExternalReference`
+
+## 1.67.2
+
+### Bugfixes
+
+- Added `Wizaplace\SDK\Pim\Product\ProductService::addAttachments`
+- Added `Wizaplace\SDK\Pim\Product\ProductService::removeAttachment`
+- Added `Wizaplace\SDK\Catalog\CatalogService::getProductAttachment`
+
+## 1.66.0
+
+### New features
+
+- Added `Wizaplace\SDK\Vendor\Order\Order::getDetails` 
+- Added `Wizaplace\SDK\Vendor\Order\OrderService::setOrderDetails`
+- Added `alt` property to `\Wizaplace\SDK\Cms\Banner`
 
 ### Bugfixes